creat author

+++ app/Http/Controllers/AuthorController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Role;
 use Spatie\Permission\Models\Permission;
+use Illuminate\Support\Facades\DB;
 use Auth;
+
 class AuthorController extends Controller
 {
     public function roleIndex(){
@@ -13,17 +15,49 @@ class AuthorController extends Controller
         return view('admin.user.listRole',compact('role'));
     }
 
-    public function permissionIndex(){
-        $permission = Permission::all();
-        return view('admin.user.listPermission',compact('permission'));
-    }
 
     public function creatRole(){
+
+        if(!Auth()->user()->hasRole('admins'))
+        {
+            return abort(403, 'Unauthorized action.');
+        }
 
         return view('admin.user.addrole');
     }
 
+    public function editRole($id){
+
+        if(!Auth()->user()->hasRole('admins'))
+        {
+            return abort(403, 'Unauthorized action.');
+        }
+
+        $roles = Role::find($id);
+        return view('admin.user.editrole',compact('roles'));
+    }
+
+    public function updateRole(Request $request,$id){
+
+        if(!Auth()->user()->hasRole('admins'))
+        {
+            return abort(403, 'Unauthorized action.');
+        }
+
+        $roles = Role::findById($id);
+        $roles->name = $request->name;
+        $roles->save();
+        return redirect()->route('role');
+    }
+
+
     public function storeRole(Request $request){
+
+        if(!Auth()->user()->hasRole('admins'))
+        {
+            return abort(403, 'Unauthorized action.');
+        }
+
 
         $roles = new Role();
 
@@ -38,7 +72,17 @@ class AuthorController extends Controller
 
     public function destroyRole($id){
 
+        if(!Auth()->user()->hasRole('admins'))
+        {
+            return abort(403, 'Unauthorized action.');
+        }
+
         $roles = Role::find($id);
+        $a = $roles->getAllPermissions();
+
+        foreach ($a as $value){
+           dump($roles->revokePermissionTo($value->name));
+        }
 
         $roles->delete();
 
@@ -48,7 +92,83 @@ class AuthorController extends Controller
 
     public function creatPermission(){
 
+        if(!Auth()->user()->hasRole('admins'))
+        {
+            return abort(403, 'Unauthorized action.');
+        }
+
         return view('admin.user.addpermission');
+    }
+
+    public function permissionIndex(){
+
+        $permission = Permission::paginate(4);
+        return view('admin.user.listPermission',compact('permission'));
+
+    }
+
+    public function editPermission($id){
+
+        if(!Auth()->user()->hasRole('admins'))
+        {
+            return abort(403, 'Unauthorized action.');
+        }
+
+        $permission = Permission::find($id);
+
+        return view('admin.user.editpermission',compact('permission'));
+    }
+
+    public function updatePermission(Request $request,$id){
+
+        if(!Auth()->user()->hasRole('admins'))
+        {
+            return abort(403, 'Unauthorized action.');
+        }
+
+
+        $permission = Permission::findById($id);
+
+        $permission->name = $request->name;
+
+        $permission->save();
+
+        return redirect()->route('permission');
+    }
+
+
+    public function storePermission(Request $request){
+
+        if(!Auth()->user()->hasRole('admins'))
+        {
+            return abort(403, 'Unauthorized action.');
+        }
+
+
+        $permission = new Permission();
+
+        $permission->name = $request->name;
+
+        $permission->save();
+
+        return redirect()->route('permission');
+
+
+    }
+
+    public function destroyPermission($id){
+
+        if(!Auth()->user()->hasRole('admins'))
+        {
+            return abort(403, 'Unauthorized action.');
+        }
+
+        $permission = Permission::find($id);
+
+        $permission->delete();
+
+        return redirect()->route('permission');
+
     }
